Move code around - split 1000+ LoC files into smaller ones
