+++ nym-gateway-probe/src/common/socks5_test/mod.rs
@@ -73,6 +73,7 @@ struct SingleHttpsTestResult {
     error: Option<String>,
 }
 
+#[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 pub struct HttpsConnectivityResult {
     /// successfully completed HTTPS request
