+++ nym-gateway-probe/src/lib.rs
@@ -27,6 +27,7 @@ use tracing::*;
 use url::Url;
 
 mod common;
+pub use common::types;
 pub mod config;
 
 use crate::common::bandwidth_helpers::{
