--- nym-node-status-api/nym-node-status-api/src/http/models.rs
@@ -1,1529 +1,0 @@
-use std::net::IpAddr;
-
-use crate::monitor::ExplorerPrettyBond;
-use cosmwasm_std::{Addr, Coin, Decimal};
-use nym_mixnet_contract_common::CoinSchema;
-use nym_node_requests::api::v1::node::models::NodeDescription;
-use nym_validator_client::{
-    client::NodeId,
-    models::{AuthenticatorDetailsV1, BinaryBuildInformationOwned, IpPacketRouterDetailsV1},
-    nym_api::SkimmedNode,
-    nym_nodes::{BasicEntryInformation, NodeRole},
-};
-use serde::{Deserialize, Serialize};
-use strum_macros::EnumString;
-use tracing::{error, instrument};
-use utoipa::ToSchema;
-
-use crate::db::models::NymNodeDataDeHelper;
-use crate::node_scraper::models::BridgeInformation;
-pub(crate) use nym_node_status_client::models::TestrunAssignment;
-
-#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
-pub struct Gateway {
-    pub gateway_identity_key: String,
-    pub bonded: bool,
-    pub performance: u8,
-    pub self_described: Option<serde_json::Value>,
-    pub explorer_pretty_bond: Option<serde_json::Value>,
-    pub description: NodeDescription,
-    pub last_probe_result: Option<serde_json::Value>,
-    pub last_probe_log: Option<String>,
-    pub last_testrun_utc: Option<String>,
-    pub last_updated_utc: String,
-    pub routing_score: f32,
-    pub config_score: u32,
-    pub bridges: Option<serde_json::Value>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct BuildInformation {
-    pub build_version: String,
-    pub commit_branch: String,
-    pub commit_sha: String,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema, EnumString)]
-#[serde(rename_all = "snake_case")]
-#[strum(serialize_all = "snake_case")]
-pub enum AsnKind {
-    Residential,
-    Other,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct Asn {
-    pub asn: String,
-    pub name: String,
-    pub domain: String,
-    pub route: String,
-    pub kind: AsnKind,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct Location {
-    pub two_letter_iso_country_code: String,
-    pub latitude: f64,
-    pub longitude: f64,
-
-    pub city: String,
-    pub region: String,
-    pub org: String,
-    pub postal: String,
-    pub timezone: String,
-
-    pub asn: Option<Asn>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema, EnumString)]
-#[serde(rename_all = "snake_case")]
-#[strum(serialize_all = "snake_case")]
-#[derive(PartialEq)]
-pub enum ScoreValue {
-    Offline,
-    Low,
-    Medium,
-    High,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct DVpnGatewayPerformance {
-    last_updated_utc: String,
-    score: ScoreValue,
-    mixnet_score: ScoreValue,
-    load: ScoreValue,
-    uptime_percentage_last_24_hours: f32,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct DVpnGateway {
-    pub identity_key: String,
-    pub name: String,
-    pub description: Option<String>,
-    pub ip_packet_router: Option<IpPacketRouterDetailsV1>,
-    pub authenticator: Option<AuthenticatorDetailsV1>,
-    pub location: Location,
-    pub last_probe: Option<DirectoryGwProbe>,
-    #[schema(value_type = Vec<String>)]
-    pub ip_addresses: Vec<IpAddr>,
-    pub mix_port: u16,
-    pub role: NodeRole,
-    pub entry: Option<BasicEntryInformation>,
-    pub bridges: Option<BridgeInformation>,
-
-    // The performance data here originates from the nym-api, and is effectively mixnet performance
-    // at the time of writing this
-    pub performance: String,
-
-    // Node performance information needed by the NymVPN UI / Explorer to show more information
-    // about the node in a user-friendly way
-    pub performance_v2: Option<DVpnGatewayPerformance>,
-
-    pub build_information: BinaryBuildInformationOwned,
-}
-
-impl DVpnGateway {
-    pub fn can_route_entry(&self) -> bool {
-        self.last_probe
-            .as_ref()
-            .map(DirectoryGwProbe::can_route_entry)
-            .unwrap_or(false)
-    }
-
-    pub fn can_route_exit(&self) -> bool {
-        self.last_probe
-            .as_ref()
-            .map(|probe| probe.can_route_exit().unwrap_or(false))
-            .unwrap_or(false)
-    }
-}
-
-/// based on
-/// https://github.com/nymtech/nym-vpn-client/blob/nym-vpn-core-v1.10.0/nym-vpn-core/crates/nym-gateway-probe/src/types.rs
-/// TODO: long term types should be moved into this repo because nym-vpn-client
-/// could pull it as a dependency and we'd have a single source of truth
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct LastProbeResult {
-    node: String,
-    used_entry: String,
-    outcome: ProbeOutcome,
-}
-
-use nym_gateway_probe::types::ProbeResult as ProbeResultLatest;
-
-impl From<ProbeResultLatest> for LastProbeResult {
-    fn from(value: ProbeResultLatest) -> Self {
-        Self {
-            node: value.node.into(),
-            used_entry: value.used_entry.into(),
-            outcome: value.outcome.into(),
-        }
-    }
-}
-
-impl LastProbeResult {
-    pub(crate) fn deserialize_with_fallback(value: serde_json::Value) -> anyhow::Result<Self> {
-        // first try matching latest struct from GW probe crate
-        let probe_result = match serde_json::from_value::<ProbeResultLatest>(value.clone()) {
-            Ok(probe_result) => probe_result.into(),
-            // as a fallback, try parsing struct from this crate
-            Err(_) => match serde_json::from_value::<Self>(value) {
-                Ok(probe_result) => probe_result,
-                Err(e) => {
-                    error!("Failed to deserialize probe result: {e}");
-                    return Err(e.into());
-                }
-            },
-        };
-
-        Ok(probe_result)
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct DirectoryGwProbe {
-    last_updated_utc: String,
-    outcome: ProbeOutcome,
-}
-
-impl DirectoryGwProbe {
-    pub fn from_outcome(outcome: ProbeOutcome, last_updated_utc: String) -> Self {
-        Self {
-            last_updated_utc,
-            outcome,
-        }
-    }
-
-    pub fn can_route_entry(&self) -> bool {
-        match &self.outcome.as_entry {
-            Entry::Tested(entry_test_result) => entry_test_result.can_route,
-            Entry::NotTested | Entry::EntryFailure => false,
-        }
-    }
-
-    pub fn can_route_exit(&self) -> Option<bool> {
-        self.outcome
-            .as_exit
-            .as_ref()
-            .map(|outcome| outcome.can_route_ip_external_v4 && outcome.can_route_ip_external_v6)
-    }
-}
-
-/// this structure is parsed on VPN API so it has some fields which must not be changed
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct ProbeOutcome {
-    pub as_entry: Entry,
-    pub as_exit: Option<Exit>,
-    pub wg: Option<WgProbeResults>,
-}
-
-use nym_gateway_probe::types::ProbeOutcome as ProbeOutcomeLatest;
-
-impl From<ProbeOutcomeLatest> for ProbeOutcome {
-    fn from(value: ProbeOutcomeLatest) -> Self {
-        Self {
-            as_entry: value.as_entry.into(),
-            as_exit: value.as_exit.map(From::from),
-            wg: value.wg.map(From::from),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
-#[serde(untagged)]
-#[allow(clippy::enum_variant_names)]
-pub enum Entry {
-    Tested(EntryTestResult),
-    NotTested,
-    EntryFailure,
-}
-
-use nym_gateway_probe::types::Entry as EntryLatest;
-
-impl From<EntryLatest> for Entry {
-    fn from(value: EntryLatest) -> Self {
-        match value {
-            EntryLatest::Tested(entry_test_result) => Self::Tested(entry_test_result.into()),
-            EntryLatest::NotTested => Self::NotTested,
-            EntryLatest::EntryFailure => Self::EntryFailure,
-        }
-    }
-}
-
-use nym_gateway_probe::types::EntryTestResult as EntryTestResultLatest;
-
-#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
-pub struct EntryTestResult {
-    pub can_connect: bool,
-    pub can_route: bool,
-}
-
-impl From<EntryTestResultLatest> for EntryTestResult {
-    fn from(value: EntryTestResultLatest) -> Self {
-        Self {
-            can_connect: value.can_connect,
-            can_route: value.can_route,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct Exit {
-    pub can_connect: bool,
-    pub can_route_ip_v4: bool,
-    pub can_route_ip_external_v4: bool,
-    pub can_route_ip_v6: bool,
-    pub can_route_ip_external_v6: bool,
-}
-
-use nym_gateway_probe::types::Exit as ExitLatest;
-
-impl From<ExitLatest> for Exit {
-    fn from(value: ExitLatest) -> Self {
-        Self {
-            can_connect: value.can_connect,
-            can_route_ip_v4: value.can_route_ip_v4,
-            can_route_ip_external_v4: value.can_route_ip_external_v4,
-            can_route_ip_v6: value.can_route_ip_v6,
-            can_route_ip_external_v6: value.can_route_ip_external_v6,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct WgProbeResults {
-    // mandatory fields
-    pub can_register: bool,
-    pub can_handshake: Option<bool>,
-    pub can_resolve_dns: Option<bool>,
-    pub ping_hosts_performance: Option<f32>,
-    pub ping_ips_performance: Option<f32>,
-
-    pub can_query_metadata_v4: Option<bool>,
-    pub can_handshake_v4: bool,
-    pub can_resolve_dns_v4: bool,
-    pub ping_hosts_performance_v4: f32,
-    pub ping_ips_performance_v4: f32,
-
-    pub can_handshake_v6: bool,
-    pub can_resolve_dns_v6: bool,
-    pub ping_hosts_performance_v6: f32,
-    pub ping_ips_performance_v6: f32,
-
-    pub download_duration_sec_v4: u64,
-    pub download_duration_milliseconds_v4: Option<u64>,
-    pub downloaded_file_size_bytes_v4: Option<u64>,
-    pub downloaded_file_v4: String,
-    pub download_error_v4: String,
-
-    pub download_duration_sec_v6: u64,
-    pub download_duration_milliseconds_v6: Option<u64>,
-    pub downloaded_file_size_bytes_v6: Option<u64>,
-    pub downloaded_file_v6: String,
-    pub download_error_v6: String,
-}
-
-use nym_gateway_probe::types::WgProbeResults as WgProbeResultsLatest;
-
-impl From<WgProbeResultsLatest> for WgProbeResults {
-    fn from(value: WgProbeResultsLatest) -> Self {
-        Self {
-            can_register: value.can_register,
-            can_handshake: Some(value.can_handshake_v4),
-            can_resolve_dns: Some(value.can_resolve_dns_v4),
-            ping_hosts_performance: Some(value.ping_hosts_performance_v4),
-            ping_ips_performance: Some(value.ping_ips_performance_v4),
-
-            can_query_metadata_v4: Some(value.can_query_metadata_v4),
-            can_handshake_v4: value.can_handshake_v4,
-            can_resolve_dns_v4: value.can_resolve_dns_v4,
-            ping_hosts_performance_v4: value.ping_hosts_performance_v4,
-            ping_ips_performance_v4: value.ping_ips_performance_v4,
-
-            can_handshake_v6: value.can_handshake_v6,
-            can_resolve_dns_v6: value.can_resolve_dns_v6,
-            ping_hosts_performance_v6: value.ping_hosts_performance_v6,
-            ping_ips_performance_v6: value.ping_ips_performance_v6,
-
-            download_duration_sec_v4: value.download_duration_sec_v4,
-            download_duration_milliseconds_v4: Some(value.download_duration_milliseconds_v4),
-            downloaded_file_size_bytes_v4: Some(value.downloaded_file_size_bytes_v4),
-            downloaded_file_v4: value.downloaded_file_v4,
-            download_error_v4: value.download_error_v4,
-
-            download_duration_sec_v6: value.download_duration_sec_v6,
-            download_duration_milliseconds_v6: Some(value.download_duration_milliseconds_v6),
-            downloaded_file_size_bytes_v6: Some(value.downloaded_file_size_bytes_v6),
-            downloaded_file_v6: value.downloaded_file_v6,
-            download_error_v6: value.download_error_v6,
-        }
-    }
-}
-
-impl DVpnGateway {
-    #[instrument(level = tracing::Level::INFO, name = "dvpn_gw_new", skip_all, fields(gateway_key = gateway.gateway_identity_key, node_id = skimmed_node.node_id))]
-    pub(crate) fn new(gateway: Gateway, skimmed_node: &SkimmedNode) -> anyhow::Result<Self> {
-        let location = gateway
-            .explorer_pretty_bond
-            .clone()
-            .ok_or_else(|| anyhow::anyhow!("Missing explorer_pretty_bond"))
-            .and_then(|value| {
-                serde_json::from_value::<ExplorerPrettyBond>(value).map_err(From::from)
-            })
-            .map(|bond| bond.location)?;
-
-        let self_described: NymNodeDataDeHelper = gateway
-            .self_described
-            .clone()
-            .ok_or_else(|| anyhow::anyhow!("Missing self_described"))
-            .and_then(|value| {
-                serde_json::from_value::<NymNodeDataDeHelper>(value).map_err(From::from)
-            })?;
-
-        let last_updated_utc = gateway.last_testrun_utc.clone().unwrap_or_default();
-        let performance = to_percent(gateway.performance);
-        let network_monitor_performance_mixnet_mode = gateway.performance as f32 / 100f32;
-        let bridges = gateway.bridges.clone().and_then(|v| {
-            serde_json::from_value(v)
-                .inspect_err(|err| {
-                    error!(
-                        "Failed to deserialize bridges for gateway identity {}: {err}",
-                        gateway.gateway_identity_key
-                    );
-                })
-                .ok()
-        });
-
-        tracing::info!("🌈 gateway probe result: {:?}", gateway.last_probe_result);
-
-        let (last_probe_result, performance_v2) = match gateway.last_probe_result {
-            Some(ref value) => {
-                let mut parsed = LastProbeResult::deserialize_with_fallback(value.clone())
-                    .inspect_err(|err| {
-                        error!("Failed to deserialize probe result: {err}");
-                    })?;
-
-                parsed.outcome.wg = parsed.outcome.wg.clone().map(|mut wg| {
-                    if wg.can_handshake.is_none() {
-                        wg.can_handshake = Some(wg.can_handshake_v4);
-                    }
-                    if wg.can_resolve_dns.is_none() {
-                        wg.can_resolve_dns = Some(wg.can_resolve_dns_v4);
-                    }
-                    if wg.ping_hosts_performance.is_none() {
-                        wg.ping_hosts_performance = Some(wg.ping_hosts_performance_v4);
-                    }
-                    if wg.ping_ips_performance.is_none() {
-                        wg.ping_ips_performance = Some(wg.ping_ips_performance_v4);
-                    }
-                    wg
-                });
-
-                tracing::info!("🌈 gateway probe parsed: {:?}", parsed);
-                let mixnet_score = calculate_mixnet_score(&gateway);
-                let score = calculate_score(&gateway, &parsed);
-                let mut load = calculate_load(&parsed);
-
-                // clamp the load value to offline, when the score is offline
-                if score == ScoreValue::Offline {
-                    load = ScoreValue::Offline;
-                }
-
-                let performance_v2 = DVpnGatewayPerformance {
-                    last_updated_utc: last_updated_utc.to_string(),
-                    load,
-                    score,
-                    mixnet_score,
-
-                    // the network monitor's measure is a good proxy for node uptime, it can be improved in the future
-                    uptime_percentage_last_24_hours: network_monitor_performance_mixnet_mode,
-                };
-                (Some(parsed), Some(performance_v2))
-            }
-            None => (None, None),
-        };
-
-        Ok(Self {
-            identity_key: gateway.gateway_identity_key,
-            name: gateway.description.moniker,
-            description: Some(gateway.description.details),
-            ip_packet_router: self_described.ip_packet_router,
-            authenticator: self_described.authenticator,
-            location: Location {
-                latitude: location.location.latitude,
-                longitude: location.location.longitude,
-                two_letter_iso_country_code: location.two_letter_iso_country_code,
-                org: location.org,
-                city: location.city,
-                region: location.region,
-                postal: location.postal,
-                timezone: location.timezone,
-                asn: location.asn.map(|a| {
-                    let kind = if a.kind.eq_ignore_ascii_case("isp") {
-                        // we consider anything that is "ISP" from ipinfo to be residential
-                        AsnKind::Residential
-                    } else {
-                        // everything else is considered "other"
-                        AsnKind::Other
-                    };
-                    Asn {
-                        asn: a.asn,
-                        domain: a.domain,
-                        kind,
-                        name: a.name,
-                        route: a.route,
-                    }
-                }),
-            },
-            last_probe: last_probe_result
-                .map(|res| DirectoryGwProbe::from_outcome(res.outcome, last_updated_utc)),
-            ip_addresses: skimmed_node.ip_addresses.clone(),
-            mix_port: skimmed_node.mix_port,
-            role: skimmed_node.role.clone(),
-            entry: skimmed_node.entry.clone(),
-            bridges,
-            performance,
-            performance_v2,
-            build_information: self_described.build_information,
-        })
-    }
-}
-
-struct NodeScore {
-    download_speed_score: f64,
-    ping_ips_score: f64,
-    mixnet_performance: f64,
-}
-
-impl NodeScore {
-    // Weighted scoring: mixnet performance (40%), download speed (30%), ping performance (30%)
-    fn calculate_weighted_score(&self) -> f64 {
-        (self.mixnet_performance * 0.4)
-            + (self.download_speed_score * 0.3)
-            + (self.ping_ips_score * 0.3)
-    }
-}
-
-/// calculates the gateway probe score for mixnet mode
-fn calculate_mixnet_score(gateway: &Gateway) -> ScoreValue {
-    let mixnet_performance = gateway.performance as f64 / 100.0;
-
-    if mixnet_performance > 0.8 {
-        ScoreValue::High
-    } else if mixnet_performance > 0.6 {
-        ScoreValue::Medium
-    } else if mixnet_performance > 0.1 {
-        ScoreValue::Low
-    } else {
-        ScoreValue::Offline
-    }
-}
-
-/// calculates a visual score for the gateway using weighted metrics
-fn calculate_score(gateway: &Gateway, probe_outcome: &LastProbeResult) -> ScoreValue {
-    let mixnet_performance = gateway.performance as f64 / 100.0;
-
-    let node_score = probe_outcome
-        .outcome
-        .wg
-        .as_ref()
-        .map(|p| {
-            let ping_ips_performance = p.ping_ips_performance_v4 as f64;
-
-            let duration_sec =
-                p.download_duration_milliseconds_v4
-                    .unwrap_or_else(|| p.download_duration_sec_v4 * 1000) as f64
-                    / 1000f64;
-
-            // get the file size downloaded in bytes and convert to MB, or default to 1MB
-            let file_size_mb =
-                p.downloaded_file_size_bytes_v4.unwrap_or(1048576) as f64 / 1024f64 / 1024f64;
-            let speed_mbps = file_size_mb / duration_sec;
-
-            let file_download_score = if speed_mbps > 5.0 {
-                1.0
-            } else if speed_mbps > 2.0 {
-                0.75
-            } else if speed_mbps > 1.0 {
-                0.5
-            } else if speed_mbps > 0.5 {
-                0.25
-            } else {
-                0.1
-            };
-
-            NodeScore {
-                download_speed_score: file_download_score,
-                ping_ips_score: ping_ips_performance,
-                mixnet_performance,
-            }
-        })
-        .unwrap_or(NodeScore {
-            download_speed_score: 0.0,
-            ping_ips_score: 0.0,
-            mixnet_performance,
-        });
-
-    let weighted_score = node_score.calculate_weighted_score();
-
-    if weighted_score > 0.75 {
-        ScoreValue::High
-    } else if weighted_score > 0.5 {
-        ScoreValue::Medium
-    } else if weighted_score > 0.1 {
-        ScoreValue::Low
-    } else {
-        ScoreValue::Offline
-    }
-}
-
-/// calculates a visual load score for the gateway
-fn calculate_load(probe_outcome: &LastProbeResult) -> ScoreValue {
-    let score = probe_outcome
-        .outcome
-        .wg
-        .clone()
-        .map(|p| p.ping_ips_performance_v4 as f64)
-        .unwrap_or(0f64);
-
-    if score > 0.8 {
-        ScoreValue::Low
-    } else if score > 0.4 {
-        ScoreValue::Medium
-    } else {
-        ScoreValue::High
-    }
-}
-
-fn to_percent(performance: u8) -> String {
-    let fraction = performance as f32 / 100.0;
-    format!("{fraction:.2}")
-}
-
-#[cfg(test)]
-mod test {
-
-    use super::*;
-    use std::str::FromStr;
-
-    #[test]
-    fn to_percent_should_work() {
-        let starting = [0u8, 33, 50, 99, 100];
-        let expected = ["0.00", "0.33", "0.50", "0.99", "1.00"];
-
-        for (starting, expected) in starting.into_iter().zip(expected) {
-            assert_eq!(expected, to_percent(starting));
-        }
-    }
-
-    #[test]
-    fn to_percent_edge_cases() {
-        // Test edge cases
-        assert_eq!("0.00", to_percent(0));
-        assert_eq!("1.00", to_percent(100));
-        assert_eq!("2.55", to_percent(255)); // Over 100%
-    }
-
-    #[test]
-    fn node_delegation_from_conversion() {
-        use cosmwasm_std::Uint128;
-
-        let delegation = nym_mixnet_contract_common::Delegation {
-            node_id: 42,
-            amount: Coin {
-                denom: "unym".to_string(),
-                amount: Uint128::new(1000000),
-            },
-            cumulative_reward_ratio: Decimal::from_str("1.23456789").unwrap(),
-            height: 12345,
-            owner: Addr::unchecked("owner1"),
-            proxy: Some(Addr::unchecked("proxy1")),
-        };
-
-        let node_delegation: NodeDelegation = delegation.clone().into();
-
-        assert_eq!(node_delegation.amount.denom, "unym");
-        assert_eq!(node_delegation.amount.amount, Uint128::new(1000000));
-        assert_eq!(node_delegation.cumulative_reward_ratio, "1.23456789");
-        assert_eq!(node_delegation.block_height, 12345);
-        assert_eq!(node_delegation.owner, Addr::unchecked("owner1"));
-        assert_eq!(node_delegation.proxy, Some(Addr::unchecked("proxy1")));
-    }
-
-    #[test]
-    fn node_delegation_from_conversion_no_proxy() {
-        use cosmwasm_std::Uint128;
-
-        let delegation = nym_mixnet_contract_common::Delegation {
-            node_id: 0,
-            amount: Coin {
-                denom: "uatom".to_string(),
-                amount: Uint128::new(0),
-            },
-            cumulative_reward_ratio: Decimal::zero(),
-            height: 0,
-            owner: Addr::unchecked("owner2"),
-            proxy: None,
-        };
-
-        let node_delegation: NodeDelegation = delegation.into();
-
-        assert_eq!(node_delegation.amount.denom, "uatom");
-        assert_eq!(node_delegation.amount.amount, Uint128::new(0));
-        assert_eq!(node_delegation.cumulative_reward_ratio, "0");
-        assert_eq!(node_delegation.block_height, 0);
-        assert_eq!(node_delegation.owner, Addr::unchecked("owner2"));
-        assert_eq!(node_delegation.proxy, None);
-    }
-
-    #[test]
-    fn node_delegation_from_conversion_max_values() {
-        use cosmwasm_std::Uint128;
-
-        let delegation = nym_mixnet_contract_common::Delegation {
-            node_id: u32::MAX,
-            amount: Coin {
-                denom: "test".to_string(),
-                amount: Uint128::MAX,
-            },
-            cumulative_reward_ratio: Decimal::from_str("999999999.999999999").unwrap(),
-            height: u64::MAX,
-            owner: Addr::unchecked("owner3"),
-            proxy: Some(Addr::unchecked("proxy3")),
-        };
-
-        let node_delegation: NodeDelegation = delegation.into();
-
-        assert_eq!(node_delegation.amount.amount, Uint128::MAX);
-        assert_eq!(
-            node_delegation.cumulative_reward_ratio,
-            "999999999.999999999"
-        );
-        assert_eq!(node_delegation.block_height, u64::MAX);
-    }
-
-    #[test]
-    fn location_struct_creation() {
-        let location = Location {
-            two_letter_iso_country_code: "US".to_string(),
-            latitude: 40.7128,
-            longitude: -74.0060,
-            org: "Nym".to_string(),
-            city: "Genève".to_string(),
-            region: "Geneva".to_string(),
-            postal: "1200".to_string(),
-            timezone: "Europe/Zurich".to_string(),
-            asn: None,
-        };
-
-        assert_eq!(location.two_letter_iso_country_code, "US");
-        assert_eq!(location.latitude, 40.7128);
-        assert_eq!(location.longitude, -74.0060);
-    }
-
-    #[test]
-    fn location_extreme_coordinates() {
-        // Test extreme coordinates
-        let north_pole = Location {
-            two_letter_iso_country_code: "XX".to_string(),
-            latitude: 90.0,
-            longitude: 0.0,
-            org: "Nym".to_string(),
-            city: "Genève".to_string(),
-            region: "Geneva".to_string(),
-            postal: "1200".to_string(),
-            timezone: "Europe/Zurich".to_string(),
-            asn: None,
-        };
-
-        let south_pole = Location {
-            two_letter_iso_country_code: "AQ".to_string(),
-            latitude: -90.0,
-            longitude: 0.0,
-            org: "Nym".to_string(),
-            city: "Genève".to_string(),
-            region: "Geneva".to_string(),
-            postal: "1200".to_string(),
-            timezone: "Europe/Zurich".to_string(),
-            asn: None,
-        };
-
-        let date_line = Location {
-            two_letter_iso_country_code: "FJ".to_string(),
-            latitude: -17.0,
-            longitude: 180.0,
-            org: "Nym".to_string(),
-            city: "Genève".to_string(),
-            region: "Geneva".to_string(),
-            postal: "1200".to_string(),
-            timezone: "Europe/Zurich".to_string(),
-            asn: None,
-        };
-
-        assert_eq!(north_pole.latitude, 90.0);
-        assert_eq!(south_pole.latitude, -90.0);
-        assert_eq!(date_line.longitude, 180.0);
-    }
-
-    #[test]
-    fn build_information_creation() {
-        let build_info = BuildInformation {
-            build_version: "1.2.3".to_string(),
-            commit_branch: "main".to_string(),
-            commit_sha: "abcdef123456".to_string(),
-        };
-
-        assert_eq!(build_info.build_version, "1.2.3");
-        assert_eq!(build_info.commit_branch, "main");
-        assert_eq!(build_info.commit_sha, "abcdef123456");
-    }
-
-    #[test]
-    fn daily_stats_creation() {
-        let stats = DailyStats {
-            date_utc: "2024-01-20".to_string(),
-            total_packets_received: 1_000_000,
-            total_packets_sent: 999_000,
-            total_packets_dropped: 1_000,
-            total_stake: 5_000_000,
-        };
-
-        assert_eq!(stats.date_utc, "2024-01-20");
-        assert_eq!(stats.total_packets_received, 1_000_000);
-        assert_eq!(stats.total_packets_sent, 999_000);
-        assert_eq!(stats.total_packets_dropped, 1_000);
-        assert_eq!(stats.total_stake, 5_000_000);
-    }
-
-    #[test]
-    fn daily_stats_negative_values() {
-        // Test with edge case values
-        let stats = DailyStats {
-            date_utc: "".to_string(),
-            total_packets_received: i64::MAX,
-            total_packets_sent: 0,
-            total_packets_dropped: -1, // Should this be allowed?
-            total_stake: i64::MIN,
-        };
-
-        assert_eq!(stats.date_utc, "");
-        assert_eq!(stats.total_packets_received, i64::MAX);
-        assert_eq!(stats.total_packets_sent, 0);
-        assert_eq!(stats.total_packets_dropped, -1);
-        assert_eq!(stats.total_stake, i64::MIN);
-    }
-
-    #[test]
-    fn gateway_skinny_creation() {
-        let gateway = GatewaySkinny {
-            gateway_identity_key: "gateway123".to_string(),
-            self_described: Some(serde_json::json!({"test": "value"})),
-            explorer_pretty_bond: None,
-            last_probe_result: Some(serde_json::json!({"status": "ok"})),
-            last_testrun_utc: Some("2024-01-20T10:00:00Z".to_string()),
-            last_updated_utc: "2024-01-20T11:00:00Z".to_string(),
-            routing_score: 0.95,
-            config_score: 100,
-            performance: 98,
-        };
-
-        assert_eq!(gateway.gateway_identity_key, "gateway123");
-        assert!(gateway.self_described.is_some());
-        assert!(gateway.explorer_pretty_bond.is_none());
-        assert_eq!(gateway.performance, 98);
-        assert_eq!(gateway.routing_score, 0.95);
-    }
-
-    #[test]
-    fn service_creation_with_all_fields() {
-        let service = Service {
-            gateway_identity_key: "gw123".to_string(),
-            last_updated_utc: "2024-01-20T10:00:00Z".to_string(),
-            routing_score: 0.85,
-            service_provider_client_id: Some("client123".to_string()),
-            ip_address: Some("192.168.1.1".to_string()),
-            hostname: Some("gateway.example.com".to_string()),
-            mixnet_websockets: Some(serde_json::json!({"port": 8080})),
-            last_successful_ping_utc: Some("2024-01-20T09:55:00Z".to_string()),
-        };
-
-        assert_eq!(service.gateway_identity_key, "gw123");
-        assert_eq!(service.routing_score, 0.85);
-        assert_eq!(service.ip_address, Some("192.168.1.1".to_string()));
-        assert_eq!(service.hostname, Some("gateway.example.com".to_string()));
-    }
-
-    #[test]
-    fn service_creation_minimal() {
-        let service = Service {
-            gateway_identity_key: "gw456".to_string(),
-            last_updated_utc: "2024-01-20T10:00:00Z".to_string(),
-            routing_score: 0.0,
-            service_provider_client_id: None,
-            ip_address: None,
-            hostname: None,
-            mixnet_websockets: None,
-            last_successful_ping_utc: None,
-        };
-
-        assert_eq!(service.gateway_identity_key, "gw456");
-        assert_eq!(service.routing_score, 0.0);
-        assert!(service.service_provider_client_id.is_none());
-        assert!(service.ip_address.is_none());
-        assert!(service.hostname.is_none());
-        assert!(service.mixnet_websockets.is_none());
-        assert!(service.last_successful_ping_utc.is_none());
-    }
-
-    #[test]
-    fn test_weighted_score_calculation() {
-        // Helper function to create a test gateway
-        fn create_test_gateway(performance: u8) -> Gateway {
-            Gateway {
-                gateway_identity_key: "test_key".to_string(),
-                bonded: true,
-                performance,
-                self_described: None,
-                explorer_pretty_bond: None,
-                description: nym_node_requests::api::v1::node::models::NodeDescription {
-                    moniker: "test".to_string(),
-                    details: "test".to_string(),
-                    security_contact: "test@example.com".to_string(),
-                    website: "https://example.com".to_string(),
-                },
-                last_probe_result: None,
-                last_probe_log: None,
-                last_testrun_utc: None,
-                last_updated_utc: "2025-10-10T00:00:00Z".to_string(),
-                routing_score: 0.0,
-                config_score: 0,
-                bridges: None,
-            }
-        }
-
-        // Helper function to create a test probe outcome
-        fn create_test_probe_outcome(
-            download_speed_mbps: f64,
-            ping_ips_performance: f32,
-        ) -> LastProbeResult {
-            let duration_sec = 1.0;
-            let file_size_mb = download_speed_mbps;
-
-            LastProbeResult {
-                node: "test_node".to_string(),
-                used_entry: "test_entry".to_string(),
-                outcome: ProbeOutcome {
-                    as_entry: Entry::Tested(EntryTestResult {
-                        can_connect: true,
-                        can_route: true,
-                    }),
-                    as_exit: None,
-                    wg: Some(WgProbeResults {
-                        can_register: true,
-                        can_handshake: Some(true),
-                        can_resolve_dns: Some(true),
-                        ping_hosts_performance: Some(ping_ips_performance),
-                        ping_ips_performance: Some(ping_ips_performance),
-                        can_query_metadata_v4: Some(true),
-                        can_handshake_v4: true,
-                        can_resolve_dns_v4: true,
-                        ping_hosts_performance_v4: ping_ips_performance,
-                        ping_ips_performance_v4: ping_ips_performance,
-                        can_handshake_v6: true,
-                        can_resolve_dns_v6: true,
-                        ping_hosts_performance_v6: ping_ips_performance,
-                        ping_ips_performance_v6: ping_ips_performance,
-                        download_duration_sec_v4: (duration_sec * 1000.0) as u64,
-                        download_duration_milliseconds_v4: Some((duration_sec * 1000.0) as u64),
-                        downloaded_file_size_bytes_v4: Some(
-                            (file_size_mb * 1024.0 * 1024.0) as u64,
-                        ),
-                        downloaded_file_v4: "test".to_string(),
-                        download_error_v4: "".to_string(),
-                        download_duration_sec_v6: 0,
-                        download_duration_milliseconds_v6: Some(0),
-                        downloaded_file_size_bytes_v6: Some(0),
-                        downloaded_file_v6: "".to_string(),
-                        download_error_v6: "".to_string(),
-                    }),
-                },
-            }
-        }
-
-        // Test case 1: Excellent node (should be High)
-        let gateway = create_test_gateway(90); // 90% mixnet performance
-        let probe = create_test_probe_outcome(6.0, 1.0); // 6 Mbps, 100% ping
-        let score = calculate_score(&gateway, &probe);
-        assert_eq!(score, ScoreValue::High, "Excellent node should be High");
-
-        // Test case 2: Good node (should be High with weighted scoring)
-        let gateway = create_test_gateway(90); // 90% mixnet performance
-        let probe = create_test_probe_outcome(3.0, 0.9); // 3 Mbps (0.75 score), 90% ping
-        let score = calculate_score(&gateway, &probe);
-        assert_eq!(
-            score,
-            ScoreValue::High,
-            "Good node should be High with weighted scoring"
-        );
-
-        // Test case 3: Medium node
-        let gateway = create_test_gateway(80); // 80% mixnet performance
-        let probe = create_test_probe_outcome(1.5, 0.8); // 1.5 Mbps (0.5 score), 80% ping
-        let score = calculate_score(&gateway, &probe);
-        assert_eq!(score, ScoreValue::Medium, "Medium node should be Medium");
-
-        // Test case 4: Poor node
-        let gateway = create_test_gateway(60); // 60% mixnet performance
-        let probe = create_test_probe_outcome(0.3, 0.3); // 0.3 Mbps (0.1 score), 30% ping
-        let score = calculate_score(&gateway, &probe);
-        assert_eq!(score, ScoreValue::Low, "Poor node should be Low");
-
-        // Test case 5: Failed node
-        let gateway = create_test_gateway(10); // 10% mixnet performance
-        let probe = create_test_probe_outcome(0.1, 0.0); // 0.1 Mbps (0.1 score), 0% ping
-        let score = calculate_score(&gateway, &probe);
-        assert_eq!(score, ScoreValue::Offline, "Failed node should be Offline");
-
-        // Test case 6: Edge case - just above threshold
-        let gateway = create_test_gateway(76); // 76% mixnet performance
-        let probe = create_test_probe_outcome(2.1, 0.75); // 2.1 Mbps (0.75 score), 75% ping
-        let score = calculate_score(&gateway, &probe);
-        // Weighted: (0.76 * 0.4) + (0.75 * 0.3) + (0.75 * 0.3) = 0.304 + 0.225 + 0.225 = 0.754
-        assert_eq!(
-            score,
-            ScoreValue::High,
-            "Edge case just above 0.75 threshold should be High"
-        );
-    }
-
-    /// Smoke test to ensure conversion from nym_gateway_probe crate's ProbeResult
-    /// to this crate's ProbeResult doesn't silently drop or default fields when
-    /// nym_gateway_probe types change.
-    ///
-    /// All values are set to non-default (booleans=true, numbers non-zero, strings non-empty)
-    /// to catch cases where new fields might be left as default after conversion.
-    #[test]
-    fn conversion_from_gw_probe_latest() {
-        use nym_gateway_probe::types::{
-            Entry as EntryLatest, EntryTestResult as EntryTestResultLatest, Exit as ExitLatest,
-            ProbeOutcome as ProbeOutcomeLatest, ProbeResult as ProbeResultLatest,
-            WgProbeResults as WgProbeResultsLatest,
-        };
-
-        // Build a ProbeResultLatest with ALL non-default values
-        let wg_latest = WgProbeResultsLatest {
-            can_register: true,
-            can_query_metadata_v4: true,
-            can_handshake_v4: true,
-            can_resolve_dns_v4: true,
-            ping_hosts_performance_v4: 0.95,
-            ping_ips_performance_v4: 0.92,
-            can_handshake_v6: true,
-            can_resolve_dns_v6: true,
-            ping_hosts_performance_v6: 0.88,
-            ping_ips_performance_v6: 0.85,
-            download_duration_sec_v4: 5,
-            download_duration_milliseconds_v4: 5123,
-            downloaded_file_size_bytes_v4: 10485760,
-            downloaded_file_v4: "test-file-v4.bin".to_string(),
-            download_error_v4: "none-v4".to_string(),
-            download_duration_sec_v6: 6,
-            download_duration_milliseconds_v6: 6234,
-            downloaded_file_size_bytes_v6: 20971520,
-            downloaded_file_v6: "test-file-v6.bin".to_string(),
-            download_error_v6: "none-v6".to_string(),
-        };
-        let probe_latest = ProbeResultLatest {
-            node: "test-node-identity-key".to_string(),
-            used_entry: "test-entry-node".to_string(),
-            outcome: ProbeOutcomeLatest {
-                as_entry: EntryLatest::Tested(EntryTestResultLatest {
-                    can_connect: true,
-                    can_route: true,
-                }),
-                as_exit: Some(ExitLatest {
-                    can_connect: true,
-                    can_route_ip_v4: true,
-                    can_route_ip_external_v4: true,
-                    can_route_ip_v6: true,
-                    can_route_ip_external_v6: true,
-                }),
-                // TODO socks5 and lp fields
-                socks5: None,
-                lp: None,
-                wg: Some(wg_latest.clone()),
-            },
-        };
-
-        // convert to this crate's LastProbeResult
-        let result: LastProbeResult = probe_latest.clone().into();
-
-        assert_eq!(result.node, probe_latest.node);
-        assert_eq!(result.used_entry, probe_latest.used_entry);
-
-        match &result.outcome.as_entry {
-            Entry::Tested(entry) => {
-                assert!(entry.can_connect);
-                assert!(entry.can_route);
-            }
-            other => panic!("Expected Entry::Tested, got {:?}", other),
-        }
-
-        // Exit conversion
-        let exit = result
-            .outcome
-            .as_exit
-            .as_ref()
-            .expect("as_exit should be Some");
-        assert!(exit.can_connect);
-        assert!(exit.can_route_ip_v4);
-        assert!(exit.can_route_ip_external_v4,);
-        assert!(exit.can_route_ip_v6);
-        assert!(exit.can_route_ip_external_v6,);
-
-        // WgProbeResults conversion
-        let wg = result.outcome.wg.as_ref().expect("wg should be Some");
-        assert!(wg.can_register);
-        assert_eq!(wg.can_query_metadata_v4, Some(true),);
-        assert!(wg.can_handshake_v4);
-        assert!(wg.can_resolve_dns_v4,);
-        assert_eq!(
-            wg.ping_hosts_performance_v4,
-            wg_latest.ping_hosts_performance_v4
-        );
-        assert_eq!(
-            wg.ping_ips_performance_v4,
-            wg_latest.ping_ips_performance_v4
-        );
-        assert!(wg.can_handshake_v6);
-        assert!(wg.can_resolve_dns_v6);
-        assert_eq!(
-            wg.ping_hosts_performance_v6,
-            wg_latest.ping_hosts_performance_v6
-        );
-        assert_eq!(
-            wg.ping_ips_performance_v6,
-            wg_latest.ping_ips_performance_v6
-        );
-        assert_eq!(
-            wg.download_duration_sec_v4,
-            wg_latest.download_duration_sec_v4
-        );
-        assert_eq!(
-            wg.download_duration_milliseconds_v4,
-            Some(wg_latest.download_duration_milliseconds_v4),
-        );
-        assert_eq!(
-            wg.downloaded_file_size_bytes_v4,
-            Some(wg_latest.downloaded_file_size_bytes_v4),
-        );
-        assert_eq!(wg.downloaded_file_v4, wg_latest.downloaded_file_v4);
-        assert_eq!(wg.download_error_v4, wg_latest.download_error_v4);
-        assert_eq!(
-            wg.download_duration_sec_v6,
-            wg_latest.download_duration_sec_v6
-        );
-        assert_eq!(
-            wg.download_duration_milliseconds_v6,
-            Some(wg_latest.download_duration_milliseconds_v6)
-        );
-        assert_eq!(
-            wg.downloaded_file_size_bytes_v6,
-            Some(wg_latest.downloaded_file_size_bytes_v6)
-        );
-        assert_eq!(wg.downloaded_file_v6, wg_latest.downloaded_file_v6);
-        assert_eq!(wg.download_error_v6, wg_latest.download_error_v6);
-
-        // fields that map from v4 values
-        assert_eq!(wg.can_handshake, Some(true));
-        assert_eq!(wg.can_resolve_dns, Some(true));
-        assert_eq!(
-            wg.ping_hosts_performance,
-            Some(wg_latest.ping_hosts_performance_v4)
-        );
-        assert_eq!(
-            wg.ping_ips_performance,
-            Some(wg_latest.ping_ips_performance_v4)
-        );
-    }
-
-    #[test]
-    fn conversion_entry_variants() {
-        use nym_gateway_probe::types::Entry as EntryLatest;
-
-        let not_tested: Entry = EntryLatest::NotTested.into();
-        assert!(matches!(not_tested, Entry::NotTested));
-
-        let failure: Entry = EntryLatest::EntryFailure.into();
-        assert!(matches!(failure, Entry::EntryFailure));
-    }
-
-    /// Backwards compatibility: this crate's struct may be present in DB of
-    /// some gateways even after the new nym_gateway_probe format is published.
-    /// DB entry needs to stay deserializable into a valid struct.
-    #[test]
-    fn deserialize_this_crate_format() {
-        // JSON that matches this crate's ProbeResult format (not nym_gateway_probe)
-        let old_format_json = serde_json::json!({
-            "node": "old-node-key",
-            "used_entry": "old-entry-key",
-            "outcome": {
-                "as_entry": {
-                    "can_connect": true,
-                    "can_route": false
-                },
-                "as_exit": null,
-                "wg": null
-            }
-        });
-
-        let result = LastProbeResult::deserialize_with_fallback(old_format_json)
-            .expect("Should deserialize old format");
-
-        assert_eq!(result.node, "old-node-key");
-        assert_eq!(result.used_entry, "old-entry-key");
-        match &result.outcome.as_entry {
-            Entry::Tested(entry) => {
-                assert!(entry.can_connect);
-                assert!(!entry.can_route);
-            }
-            other => panic!("Expected Entry::Tested, got {:?}", other),
-        }
-        assert!(result.outcome.as_exit.is_none());
-        assert!(result.outcome.wg.is_none());
-    }
-
-    /// Test that the latest nym_gateway_probe format deserializes correctly
-    #[test]
-    fn deserialize_latest_gw_probe_format() {
-        // JSON that matches nym_gateway_probe::types::ProbeResult format
-        let latest_format_json = serde_json::json!({
-            "node": "new-node-key",
-            "used_entry": "new-entry-key",
-            "outcome": {
-                "as_entry": {
-                    "can_connect": true,
-                    "can_route": true
-                },
-                "as_exit": {
-                    "can_connect": true,
-                    "can_route_ip_v4": true,
-                    "can_route_ip_external_v4": true,
-                    "can_route_ip_v6": false,
-                    "can_route_ip_external_v6": false
-                },
-                "socks5": null,
-                "lp": null,
-                "wg": {
-                    "can_register": true,
-                    "can_query_metadata_v4": true,
-                    "can_handshake_v4": true,
-                    "can_resolve_dns_v4": true,
-                    "ping_hosts_performance_v4": 0.9,
-                    "ping_ips_performance_v4": 0.85,
-                    "can_handshake_v6": false,
-                    "can_resolve_dns_v6": false,
-                    "ping_hosts_performance_v6": 0.0,
-                    "ping_ips_performance_v6": 0.0,
-                    "download_duration_sec_v4": 3,
-                    "download_duration_milliseconds_v4": 3456,
-                    "downloaded_file_size_bytes_v4": 5242880,
-                    "downloaded_file_v4": "5mb.bin",
-                    "download_error_v4": "",
-                    "download_duration_sec_v6": 0,
-                    "download_duration_milliseconds_v6": 0,
-                    "downloaded_file_size_bytes_v6": 0,
-                    "downloaded_file_v6": "",
-                    "download_error_v6": "ipv6 not supported"
-                }
-            }
-        });
-
-        let result = LastProbeResult::deserialize_with_fallback(latest_format_json)
-            .expect("Should deserialize latest format");
-
-        assert_eq!(result.node, "new-node-key");
-        assert_eq!(result.used_entry, "new-entry-key");
-
-        let exit = result.outcome.as_exit.as_ref().expect("should have exit");
-        assert!(exit.can_route_ip_external_v4);
-        assert!(!exit.can_route_ip_external_v6);
-
-        let wg = result.outcome.wg.as_ref().expect("should have wg");
-        assert!(wg.can_register);
-        assert_eq!(wg.download_duration_milliseconds_v4, Some(3456));
-        assert_eq!(wg.download_error_v6, "ipv6 not supported");
-    }
-
-    /// Serialize LastProbeResult to JSON and back to ensure serde attributes
-    /// work correctly and no data is lost.
-    #[test]
-    fn round_trip_serialization() {
-        let original = LastProbeResult {
-            node: "round-trip-node".to_string(),
-            used_entry: "round-trip-entry".to_string(),
-            outcome: ProbeOutcome {
-                as_entry: Entry::Tested(EntryTestResult {
-                    can_connect: true,
-                    can_route: true,
-                }),
-                as_exit: Some(Exit {
-                    can_connect: true,
-                    can_route_ip_v4: true,
-                    can_route_ip_external_v4: true,
-                    can_route_ip_v6: true,
-                    can_route_ip_external_v6: true,
-                }),
-                wg: Some(WgProbeResults {
-                    can_register: true,
-                    can_handshake: Some(true),
-                    can_resolve_dns: Some(true),
-                    ping_hosts_performance: Some(0.95),
-                    ping_ips_performance: Some(0.92),
-                    can_query_metadata_v4: Some(true),
-                    can_handshake_v4: true,
-                    can_resolve_dns_v4: true,
-                    ping_hosts_performance_v4: 0.95,
-                    ping_ips_performance_v4: 0.92,
-                    can_handshake_v6: true,
-                    can_resolve_dns_v6: true,
-                    ping_hosts_performance_v6: 0.88,
-                    ping_ips_performance_v6: 0.85,
-                    download_duration_sec_v4: 5,
-                    download_duration_milliseconds_v4: Some(5123),
-                    downloaded_file_size_bytes_v4: Some(10485760),
-                    downloaded_file_v4: "test-file-v4.bin".to_string(),
-                    download_error_v4: "none-v4".to_string(),
-                    download_duration_sec_v6: 6,
-                    download_duration_milliseconds_v6: Some(6234),
-                    downloaded_file_size_bytes_v6: Some(20971520),
-                    downloaded_file_v6: "test-file-v6.bin".to_string(),
-                    download_error_v6: "none-v6".to_string(),
-                }),
-            },
-        };
-
-        // Serialize to JSON
-        let json_string =
-            serde_json::to_string(&original).expect("Should serialize LastProbeResult to JSON");
-
-        // Deserialize back
-        let deserialized: LastProbeResult =
-            serde_json::from_str(&json_string).expect("Should deserialize JSON to LastProbeResult");
-
-        // Verify top-level fields
-        assert_eq!(deserialized.node, original.node);
-        assert_eq!(deserialized.used_entry, original.used_entry);
-
-        // Verify Entry
-        match (&original.outcome.as_entry, &deserialized.outcome.as_entry) {
-            (Entry::Tested(orig), Entry::Tested(deser)) => {
-                assert_eq!(orig.can_connect, deser.can_connect);
-                assert_eq!(orig.can_route, deser.can_route);
-            }
-            _ => panic!("Entry mismatch after round-trip"),
-        }
-
-        // Verify Exit
-        let orig_exit = original.outcome.as_exit.as_ref().unwrap();
-        let deser_exit = deserialized.outcome.as_exit.as_ref().unwrap();
-        assert_eq!(orig_exit.can_connect, deser_exit.can_connect);
-        assert_eq!(orig_exit.can_route_ip_v4, deser_exit.can_route_ip_v4);
-        assert_eq!(
-            orig_exit.can_route_ip_external_v4,
-            deser_exit.can_route_ip_external_v4
-        );
-        assert_eq!(orig_exit.can_route_ip_v6, deser_exit.can_route_ip_v6);
-        assert_eq!(
-            orig_exit.can_route_ip_external_v6,
-            deser_exit.can_route_ip_external_v6
-        );
-
-        // Verify WgProbeResults
-        let orig_wg = original.outcome.wg.as_ref().unwrap();
-        let deser_wg = deserialized.outcome.wg.as_ref().unwrap();
-        assert_eq!(orig_wg.can_register, deser_wg.can_register);
-        assert_eq!(orig_wg.can_handshake, deser_wg.can_handshake);
-        assert_eq!(orig_wg.can_resolve_dns, deser_wg.can_resolve_dns);
-        assert_eq!(
-            orig_wg.ping_hosts_performance,
-            deser_wg.ping_hosts_performance
-        );
-        assert_eq!(orig_wg.ping_ips_performance, deser_wg.ping_ips_performance);
-        assert_eq!(
-            orig_wg.can_query_metadata_v4,
-            deser_wg.can_query_metadata_v4
-        );
-        assert_eq!(orig_wg.can_handshake_v4, deser_wg.can_handshake_v4);
-        assert_eq!(orig_wg.can_resolve_dns_v4, deser_wg.can_resolve_dns_v4);
-        assert_eq!(
-            orig_wg.ping_hosts_performance_v4,
-            deser_wg.ping_hosts_performance_v4
-        );
-        assert_eq!(
-            orig_wg.ping_ips_performance_v4,
-            deser_wg.ping_ips_performance_v4
-        );
-        assert_eq!(orig_wg.can_handshake_v6, deser_wg.can_handshake_v6);
-        assert_eq!(orig_wg.can_resolve_dns_v6, deser_wg.can_resolve_dns_v6);
-        assert_eq!(
-            orig_wg.ping_hosts_performance_v6,
-            deser_wg.ping_hosts_performance_v6
-        );
-        assert_eq!(
-            orig_wg.ping_ips_performance_v6,
-            deser_wg.ping_ips_performance_v6
-        );
-        assert_eq!(
-            orig_wg.download_duration_sec_v4,
-            deser_wg.download_duration_sec_v4
-        );
-        assert_eq!(
-            orig_wg.download_duration_milliseconds_v4,
-            deser_wg.download_duration_milliseconds_v4
-        );
-        assert_eq!(
-            orig_wg.downloaded_file_size_bytes_v4,
-            deser_wg.downloaded_file_size_bytes_v4
-        );
-        assert_eq!(orig_wg.downloaded_file_v4, deser_wg.downloaded_file_v4);
-        assert_eq!(orig_wg.download_error_v4, deser_wg.download_error_v4);
-        assert_eq!(
-            orig_wg.download_duration_sec_v6,
-            deser_wg.download_duration_sec_v6
-        );
-        assert_eq!(
-            orig_wg.download_duration_milliseconds_v6,
-            deser_wg.download_duration_milliseconds_v6
-        );
-        assert_eq!(
-            orig_wg.downloaded_file_size_bytes_v6,
-            deser_wg.downloaded_file_size_bytes_v6
-        );
-        assert_eq!(orig_wg.downloaded_file_v6, deser_wg.downloaded_file_v6);
-        assert_eq!(orig_wg.download_error_v6, deser_wg.download_error_v6);
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct GatewaySkinny {
-    pub gateway_identity_key: String,
-    pub self_described: Option<serde_json::Value>,
-    pub explorer_pretty_bond: Option<serde_json::Value>,
-    pub last_probe_result: Option<serde_json::Value>,
-    pub last_testrun_utc: Option<String>,
-    pub last_updated_utc: String,
-    pub routing_score: f32,
-    pub config_score: u32,
-    pub performance: u8,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct Mixnode {
-    pub mix_id: u32,
-    pub bonded: bool,
-    pub is_dp_delegatee: bool,
-    pub total_stake: i64,
-    pub full_details: Option<serde_json::Value>,
-    pub self_described: Option<serde_json::Value>,
-    pub description: NodeDescription,
-    pub last_updated_utc: String,
-}
-
-#[derive(Clone, Debug, utoipa::ToSchema, Deserialize, Serialize)]
-pub(crate) struct ExtendedNymNode {
-    pub(crate) node_id: NodeId,
-    pub(crate) identity_key: String,
-    pub(crate) uptime: f64,
-    #[schema(value_type = String)]
-    pub(crate) total_stake: Decimal,
-    pub(crate) original_pledge: u128,
-    pub(crate) bonding_address: Option<String>,
-    pub(crate) bonded: bool,
-    pub(crate) node_type: nym_validator_client::models::DescribedNodeTypeV1,
-    pub(crate) ip_address: String,
-    pub(crate) accepted_tnc: bool,
-    pub(crate) self_description: nym_validator_client::models::NymNodeDataV2,
-    pub(crate) rewarding_details: Option<nym_mixnet_contract_common::NodeRewarding>,
-    pub(crate) description: NodeDescription,
-    pub(crate) geoip: Option<NodeGeoData>,
-}
-
-#[derive(Clone, Debug, utoipa::ToSchema, Deserialize, Serialize)]
-pub(crate) struct NodeGeoData {
-    pub(crate) city: String,
-    pub(crate) country: String,
-    pub(crate) ip_address: String,
-    pub(crate) latitude: String,
-    pub(crate) longitude: String,
-    pub(crate) org: String,
-    pub(crate) postal: String,
-    pub(crate) region: String,
-    pub(crate) timezone: String,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct DailyStats {
-    pub date_utc: String,
-    pub total_packets_received: i64,
-    pub total_packets_sent: i64,
-    pub total_packets_dropped: i64,
-    pub total_stake: i64,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct Service {
-    pub gateway_identity_key: String,
-    pub last_updated_utc: String,
-    pub routing_score: f32,
-    pub service_provider_client_id: Option<String>,
-    pub ip_address: Option<String>,
-    pub hostname: Option<String>,
-    pub mixnet_websockets: Option<serde_json::Value>,
-    pub last_successful_ping_utc: Option<String>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub(crate) struct SummaryHistory {
-    pub date: String,
-    pub value_json: serde_json::Value,
-    pub timestamp_utc: String,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct SessionStats {
-    pub gateway_identity_key: String,
-    pub node_id: u32,
-    #[serde(with = "nym_serde_helpers::date")]
-    pub day: time::Date,
-    pub unique_active_clients: i64,
-    pub session_started: i64,
-    pub users_hashes: Option<serde_json::Value>,
-    pub vpn_sessions: Option<serde_json::Value>,
-    pub mixnet_sessions: Option<serde_json::Value>,
-    pub unknown_sessions: Option<serde_json::Value>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-pub struct NodeDelegation {
-    #[schema(value_type = CoinSchema)]
-    pub amount: Coin,
-    pub cumulative_reward_ratio: String,
-    pub block_height: u64,
-    #[schema(value_type = String)]
-    pub owner: Addr,
-    #[schema(value_type = Option<String>)]
-    pub proxy: Option<Addr>,
-}
-
-impl From<nym_mixnet_contract_common::Delegation> for NodeDelegation {
-    fn from(value: nym_mixnet_contract_common::Delegation) -> Self {
-        Self {
-            amount: value.amount,
-            cumulative_reward_ratio: value.cumulative_reward_ratio.to_string(),
-            block_height: value.height,
-            owner: value.owner,
-            proxy: value.proxy,
-        }
-    }
-}
